+++ ray.go
@@ -12,6 +12,6 @@ func (r ray) direction() vec3 {
 	return r.b
 }
 
-func (r ray) pointAtParameter(t float32) vec3 {
+func (r ray) pointAtParameter(t float64) vec3 {
 	return r.a.plus(r.b.multiplyBy(t))
 }
